requirements.txt

Django==4.2.13
django-filter==24.2
djangorestframework==3.15.1
drf-yasg==1.21.7

